Stable 2D DGPE solver
Checked and tested 2D solver for DGPE
