+++ GPE_inst.py
@@ -1,5 +1,5 @@
 import numpy as np
-from GPElib import InstabilityGenerator
+from GPElib.instability_generator import InstabilityGenerator
 from GPElib.visualisation import Visualisation
 import matplotlib
 print matplotlib.matplotlib_fname()
@@ -11,7 +11,7 @@ sys.stderr = sys.stdout
 def init_instability(inst, traj_seed):
 	inst.generate_init('random', traj_seed, 10.)
 	delta = (2. * np.sqrt(1.0 * inst.N_part/inst.N_wells)) * np.random.rand()
-	x0, y0, err = inst.E_const_perturbation_XY(inst.X[0,:], inst.Y[0,:], delta)
+	x0, y0, err = inst.E_const_perturbation_XY(inst.X[:,:,0], inst.Y[:,:,0], delta)
 	x1, y1 = inst.constant_perturbation_XY(x0,y0)
 	inst.set_init_XY(x0,y0,x1,y1)
 	return err
@@ -25,7 +25,7 @@ if len(sys.argv) > 4:
 	unique_id = sys.argv[4]
 else:
 	seed_from = 0
-	seed_to = 5
+	seed_to = 3
 	my_id = 0
 	unique_id = 'ID_not_stated'
 
@@ -36,12 +36,14 @@ time = 100.
 # step = 0.00015625
 step = 0.01
 N_wells = 100
-W = 4.
-np.random.seed(78)
-e_disorder = -W  + 2 * W * np.random.rand(N_wells)
-inst = InstabilityGenerator(N_part_per_well=100, N_wells=N_wells, disorder=e_disorder, time=time, step=step,
-                            perturb_hamiltonian=False,
-                            error_J=0, error_beta=0, error_disorder=0)
+W = 0.
+
+inst = InstabilityGenerator(N_part_per_well=100,
+                             N_wells=(10,10), dimensionality=2,
+                             disorder_seed=53, time=time, step=step,
+                             perturb_hamiltonian=False,
+                             error_J=0, error_beta=0, error_disorder=0)
+
 grname = 'Instability_' + unique_id
 vis = Visualisation(is_local=0, GROUP_NAMES=grname)
 
@@ -55,6 +57,11 @@ polarisation = []
 polarisation1 = []
 dist = []
 energy = []
+
+all_x = {}
+all_y = {}
+all_x1 = {}
+all_y1 = {}
 
 for ii in needed_trajs:
 	inst.traj_seed = ii
@@ -70,6 +77,16 @@ for ii in needed_trajs:
 	polarisation1.append(inst.polarisation1)
 	dist.append(inst.distance)
 	energy.append(inst.energy)
+	all_x[ii] = inst.X
+	all_y[ii] = inst.Y
+	all_x1[ii] = inst.X1
+	all_y1[ii] = inst.Y1
+
+	# np.savez(vis.filename(my_id) + '_traj_' + str(ii),
+	# 	         step=inst.step, time=inst.time, n_steps=inst.n_steps,
+	# 	         error_code=inst.error_code, checksum=inst.consistency_checksum,
+	# 	         distance=inst.distance,
+	# 	         x=inst.X, y=inst.Y, x1=inst.X1, y1=inst.Y1)
 
 T = np.linspace(0, inst.time, inst.n_steps)
 for ii in xrange(needed_trajs.shape[0]):
@@ -78,10 +95,12 @@ for ii in xrange(needed_trajs.shape[0]):
 plt.savefig(vis.HOMEDIR + 'pics/Inst_' + unique_id + '_' + str(my_id)+'.png', format='png', dpi=100)
 
 np.savez(vis.filename(my_id),
+         x=all_x, y=all_y, x1=all_x1, y1=all_y1,
          data=answers, lambdas=lambdas, lambdas_no_regr=lambdas_no_regr,
          polar=polarisation, polar1=polarisation1,
          distance=dist, energy=energy,
          step=inst.step, time=inst.time, n_steps=inst.n_steps,
+         well_indices=inst.wells_indices,
          my_info=[seed_from, seed_to, my_id],
          needed_trajs=needed_trajs,
          description='Loschmidt echo for 10000 replicates, trying to calculate F(t)')
