+++ GPE_lyap.py
@@ -1,5 +1,5 @@
 import numpy as np
-from GPElib.LyapunovGenerator import LyapunovGenerator
+from GPElib.lyapunov_generator import LyapunovGenerator
 from GPElib.visualisation import Visualisation
 import matplotlib
 print matplotlib.matplotlib_fname()
@@ -11,7 +11,7 @@ sys.stderr = sys.stdout
 def init_instability(inst, traj_seed):
 	inst.generate_init('random', traj_seed, 10.)
 	delta = (2. * np.sqrt(1.0 * inst.N_part/inst.N_wells)) * np.random.rand()
-	x0, y0, err = inst.E_const_perturbation_XY(inst.X[0,:], inst.Y[0,:], delta)
+	x0, y0, err = inst.E_const_perturbation_XY(inst.X[:,:,0], inst.Y[:,:,0], delta)
 	x1, y1 = inst.constant_perturbation_XY(x0,y0)
 	inst.set_init_XY(x0,y0,x1,y1)
 	return err
@@ -25,7 +25,7 @@ if len(sys.argv) > 4:
 	unique_id = sys.argv[4]
 else:
 	seed_from = 0
-	seed_to = 5
+	seed_to = 1
 	needed_trajs = np.arange(seed_from, seed_to)
 	my_id = 0
 	unique_id = 'ID_not_stated'
@@ -37,12 +37,15 @@ time = 4 * 800.
 # time = 1.
 # time = 100. * 15
 # step = 0.00015625
-step = 0.01
-N_wells = 20
-W = 4.
-np.random.seed(78)
-e_disorder = -W  + 2. * W * np.random.rand(N_wells)
-lyap = LyapunovGenerator(N_part_per_well=100, N_wells=N_wells, disorder=e_disorder, reset_steps_duration=3000, time=time, step=step)
+step = 0.007
+N_wells = 100
+W = 0.
+
+lyap = LyapunovGenerator(N_part_per_well=100,
+                         W=W, disorder_seed=53,
+                         N_wells=(100,1), dimensionality=2,
+                         reset_steps_duration=3000, time=time, step=step)
+
 grname = 'GPE_lyap_' + unique_id
 vis = Visualisation(is_local=0, GROUP_NAMES=grname)
 
@@ -84,6 +87,24 @@ for i_traj, traj_seed in enumerate(needed_trajs):
 		print lyap.lambdas_no_regr
 		num_good += 1
 		plt.semilogy(lyap.T, lyap.distance)
+		np.savez(vis.filename(my_id) + '_traj_' + str(i_traj),
+		         step=lyap.step, time=lyap.time,
+		         traj_seed=lyap.traj_seed,
+		         pert_seed=lyap.pert_seed,
+		         disorder_seed=lyap.disorder_seed,
+		         n_steps=lyap.n_steps,
+		         wells_indices=lyap.wells_indices,
+		         beta=lyap.beta, W=lyap.W,
+		         J=lyap.J, N_tuple=lyap.N_tuple,
+		         energy=lyap.energy, number_of_particles=lyap.number_of_particles,
+		         eff_nonl=lyap.effective_nonlinearity,
+		         error_code=lyap.error_code, checksum=lyap.consistency_checksum,
+		         distance=lyap.distance,
+		         x=lyap.X, y=lyap.Y, x1=lyap.X1, y1=lyap.Y1,
+		         lambdas=lyap.lambdas, lambdas_no_regr=lyap.lambdas_no_regr,
+		         hist2d=lyap.histograms, hist1d=lyap.rho_histograms,
+		         hist2d1=lyap.histograms1, hist1d1=lyap.rho_histograms1)
+
 plt.savefig(vis.HOMEDIR + 'pics/Lyap_' + unique_id + '_' + str(my_id)+'.png', format='png', dpi=100)
 
 print "Error code: ", lyap.error_code
@@ -94,6 +115,6 @@ np.savez(vis.filename(my_id),
          eff_nonl=effective_nonlinearity,
          numb_of_part=numb_of_part, energies=energies,
          chosen=chosen_trajs, step=lyap.step, time=lyap.time, n_steps=lyap.n_steps,
-         my_info=[seed_from, seed_to, needed_trajs, my_id],
+         my_info=[seed_from, seed_to, my_id], needed_trajs=needed_trajs,
          checksum=lyap.consistency_checksum, error_code=lyap.error_code,
          distance=lyap.distance)
